+++ packages/@react-spectrum/actionbar/stories/ActionBar.stories.tsx
@@ -11,111 +11,24 @@
  */
 
 import {action} from '@storybook/addon-actions';
-import {ActionBar, ActionBarContainer, Item} from '../';
-import {Cell, Column, Row, TableBody, TableHeader, TableView} from '@react-spectrum/table';
-import Copy from '@spectrum-icons/workflow/Copy';
-import Delete from '@spectrum-icons/workflow/Delete';
-import Duplicate from '@spectrum-icons/workflow/Duplicate';
-import Edit from '@spectrum-icons/workflow/Edit';
-import Move from '@spectrum-icons/workflow/Move';
-import React, {useState} from 'react';
-import {Selection} from '@react-types/shared';
+import {Example} from './Example';
+import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {Text} from '@react-spectrum/text';
 import {useViewportSize} from '@react-aria/utils';
-
-let columns = [
-  {name: 'Foo', key: 'foo'},
-  {name: 'Bar', key: 'bar'},
-  {name: 'Baz', key: 'baz'}
-];
-
-let items = [
-  {test: 'Test 1', foo: 'Foo 1', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 2', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'},
-  {test: 'Test 1', foo: 'Foo 3', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 4', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'},
-  {test: 'Test 1', foo: 'Foo 5', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 6', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'},
-  {test: 'Test 1', foo: 'Foo 7', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 8', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'},
-  {test: 'Test 1', foo: 'Foo 9', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 10', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'},
-  {test: 'Test 1', foo: 'Foo 11', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 12', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'},
-  {test: 'Test 1', foo: 'Foo 13', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 14', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'},
-  {test: 'Test 1', foo: 'Foo 15', bar: 'Bar 1', yay: 'Yay 1', baz: 'Baz 1'},
-  {test: 'Test 2', foo: 'Foo 16', bar: 'Bar 2', yay: 'Yay 2', baz: 'Baz 2'}
-];
 
 storiesOf('ActionBar', module)
   .add(
     'default',
-    () => <Example />
+    () => <Example onAction={action('onAction')} />
   )
   .add(
     'isEmphasized',
-    () => <Example isEmphasized />
+    () => <Example isEmphasized onAction={action('onAction')} />
   )
   .add(
     'full width',
     () => {
       let viewport = useViewportSize();
-      return <Example isEmphasized tableWidth="100vw" containerHeight={viewport.height} isQuiet />;
+      return <Example isEmphasized tableWidth="100vw" containerHeight={viewport.height} isQuiet onAction={action('onAction')} />;
     }
   );
-
-function Example(props: any = {}) {
-  const [selectedKeys, setSelectedKeys] = useState<Selection>(new Set());
-  return (
-    <ActionBarContainer height={props.containerHeight || 300}>
-      <TableView
-        aria-label="Table"
-        isQuiet={props.isQuiet}
-        width={props.tableWidth}
-        selectedKeys={selectedKeys}
-        selectionMode="multiple"
-        onSelectionChange={(keys) => setSelectedKeys(keys)}>
-        <TableHeader columns={columns}>
-          {column => <Column>{column.name}</Column>}
-        </TableHeader>
-        <TableBody items={items}>
-          {item =>
-            (<Row key={item.foo}>
-              {key => <Cell>{item[key]}</Cell>}
-            </Row>)
-          }
-        </TableBody>
-      </TableView>
-      <ActionBar
-        selectedItemCount={selectedKeys === 'all' ? selectedKeys : selectedKeys.size}
-        onClearSelection={() => {
-          setSelectedKeys(new Set());
-        }}
-        onAction={action('onAction')}
-        {...props}>
-        <Item key="edit">
-          <Edit />
-          <Text>Edit</Text>
-        </Item>
-        <Item key="copy">
-          <Copy />
-          <Text>Copy</Text>
-        </Item>
-        <Item key="delete">
-          <Delete />
-          <Text>Delete</Text>
-        </Item>
-        <Item key="move">
-          <Move />
-          <Text>Move</Text>
-        </Item>
-        <Item key="duplicate">
-          <Duplicate />
-          <Text>Duplicate</Text>
-        </Item>
-      </ActionBar>
-    </ActionBarContainer>
-  );
-}
